Test-slim Phase 0: the safety net — golden-loop e2e + assistant pipeline tests
- backend/tests/test_assistant_pipeline.py: assistant cases extracted
  from test_smoke_evals.py (which keeps evals 1-3), plus the three
  missing branches: parse-failure fallback, tool-error path, empty
  tool registry. 17 tests.
- frontend/e2e/golden-loop.spec.ts: signup → seeded Khan matter →
  chat default surface → deterministic summary + sources → skill run
  → sign-off through the UI → output.signed with signer_is_author.
  Sign step is a hard assertion chain; killing the sign-off route
  fails the spec. Divergence documented in the spec header: the sign
  leg runs on a stub-echo matter because Khan seeds a keyed model
  (that key-missing IS the deterministic fallback under test).
- e2e fixtures: grantCapability + uploadTextDocument; two
  data-testids on the chat composer (test plumbing only).

Per docs/handovers/TEST_SLIM_ORDER_2026-06-12.md Phase 0.
